abandon namespace package
mbapi is not my stuff anymore.  it's also not beta anymore;
consider the current state of it the first and last version
there will be.

(well, i *could* consider updating it... but i don't study
at a school that uses managebac anymore.  so in a sense i can't
update even if i wanted to. :)